requirements.txt
discord.py>=2.7.1
motor>=3.7.1
pydantic>=2.12.5
python-dotenv>=1.2.2
pydantic_settings>=2.13.1
loguru==0.7.3
